+++ src/components/home.js
@@ -3,22 +3,37 @@ import { ThemeContext } from '../master/theme';
 import Glide from '@glidejs/glide';
 import '@glidejs/glide/dist/css/glide.core.min.css';
 
+// Import Image Path
 import imageMap from '../master/imagepath';
+// Import Character Information
+import characterInfo from '../master/characterInfo';
 
 const Home = () => {
+  // Set the Color Pallete
   const { theme } = useContext(ThemeContext);
+
+  // Set the Character Info
+  const character = characterInfo[theme];
 
   useEffect(() => {
     const glide = new Glide('.glide', {
       type: 'carousel',
       startAt: 0,
-      perView: 1,
+      perView: 2,
       gap: 0,
       autoplay: 5000,
       hoverpause: true,
+      keyboard: true,
       animationDuration: 1000,
       animationTimingFunc: 'ease-in-out',
-    });
+      touchRatio: 0.5,
+      breakpoints: {
+        768: {
+          perView: 1
+        }
+      }
+    }
+  );
 
     glide.mount();
 
@@ -28,31 +43,33 @@ const Home = () => {
 
   const images = imageMap[theme] || [];
 
-
   return (
-    <>
-      <div className='z-0 h-screen w-full'>
-        <div className='h-3/4 w-full flex bg-red-300'>
-          <div className="glide">
-            <div className="glide__track" data-glide-el="track">
-              <ul className="glide__slides">
-                {images.map((src, i) => (
-                  <li key={i} className="glide__slide">
-                    <img src={src} alt={`Slide ${i + 1}`} className="w-full h-full object-cover" />
-                  </li>
-                ))}
-              </ul>
-            </div>
-            <div data-glide-el="controls">
-              <button data-glide-dir="<<">Start</button>
-              <button data-glide-dir=">>">End</button>
-            </div>
+    <div className="min-h-screen w-full flex flex-col">
+      <div className="w-full bg-slate-50 pt-12">
+        <div className="glide">
+          <div className="glide__track" data-glide-el="track">
+            <ul className="glide__slides">
+              {images.map((src, i) => (
+                <li key={i} className="glide__slide">
+                  <img
+                    src={src}
+                    alt={`Slide ${i + 1}`}
+                    className="w-full h-auto object-cover rounded-md brightness-110 contrast-100 saturate-150 cursor-grab"
+                  />
+                </li>
+              ))}
+            </ul>
           </div>
         </div>
-        <div className='h-1/4 w-full bg-orange-300'></div>
       </div>
-    </>
+      <div className="flex-1 w-full px-2 sm:px-8 py-4 flex bg-slate-50">
+        <div className={`w-full p-8 text-center text-black text-lg border-t-8 bg-chisato bg-opacity-5 border-${theme} rounded-lg`}>
+          <h1 className='ff-goldman text-4xl sm:text-6xl text-slate-950'>{ character.name }</h1>
+        </div>
+      </div>
+    </div>
   );
+  
 };
 
 export default Home;