+++ src/master/characterInfo.js
@@ -1,0 +1,20 @@
+import chisatoSymbol from '../assets/symbols/chisato-icon-symbol.png'
+import takinaSymbol from '../assets/symbols/takina-icon-symbol.png'
+import kurumiSymbol from '../assets/symbols/kurumi-icon-symbol.png'
+
+const characterInfo = {
+    chisato: {
+        name: "Chisato Nishikigi",
+        symbol: chisatoSymbol,
+    },
+    takina: {
+        name: "Takina Inoue",
+        symbol: takinaSymbol,
+    },
+    kurumi: {
+        name: "Kurumi",
+        symbol: kurumiSymbol,
+    }
+}
+
+export default characterInfo;